Frogustus Animations
Applied the animations for frogustus in unity and ensured they worked, however they only work with a scale factor of 1 with USE FILE SCALE Selected on  if you want to change the scaling of them make sure to turn off the USE FILE SCALE option in the skeletal mesh and the animation fbx's model tab. Also deleted the new skeletalmesh.